requirements.txt
streamlit 
rembg 
diffusers 
transformers 
pillow 
torch 
numpy